COM-1899 - merge style
