+++ src/components/CreateAccountForm/styles.ts
@@ -1,28 +1,9 @@
 import { StyleSheet } from 'react-native';
 import { MARGIN } from '../../styles/metrics';
-import { FIRA_SANS_BOLD, FIRA_SANS_ITALIC } from '../../styles/fonts';
+import { FIRA_SANS_ITALIC } from '../../styles/fonts';
 import { GREY } from '../../styles/colors';
 
-const styles = StyleSheet.create({
-  title: {
-    ...FIRA_SANS_BOLD.LG,
-    marginVertical: MARGIN.LG,
-  },
-  keyboardAvoidingView: {
-    flexGrow: 1,
-  },
-  input: {
-    marginBottom: MARGIN.XS,
-  },
-  container: {
-    flexGrow: 1,
-    marginHorizontal: MARGIN.LG,
-  },
-  footer: {
-    justifyContent: 'flex-end',
-    marginBottom: MARGIN.MD,
-    flex: 1,
-  },
+export default StyleSheet.create({
   modalText: {
     ...FIRA_SANS_ITALIC.SM,
     color: GREY[600],
@@ -32,5 +13,3 @@ const styles = StyleSheet.create({
     textDecorationLine: 'underline',
   },
 });
-
-export default styles;
